Fix finding "Default" section.
The algorithm searches for the first occurences of
"Default" which is expected as first word in the profile file.

For the available firmware the first line is:
which contains also the word default and confuses the
algorithm.

Seach for "default\n" as workaround to return the correct
start for the current section.