WORKAROUND: Fix access to firmware data.
After some stress test, the driver crashes with an illegal access:
CPU 1 Unable to handle kernel paging request at virtual address c0ef3000, epc == 805abb28, ra == 809e0324
when accessing the fw->data, which should hold the loaded profile.

Currently it's not clear why this is the case, but for now
loading the data in a separate buffer seems to fix this issue.